+++ src/types.d.ts
@@ -3,3 +3,11 @@ declare module "*.css" {
   const content: string;
   export default content;
 }
+
+import "react";
+
+declare module "react" {
+  interface CSSProperties {
+    "--bars-height"?: string;
+  }
+}